Add command line arguments
I refactored the code into a class to more easily parse and store state
from command line arguments.

I added a verbose mode and defaulted it to off (thereby changing the
default output).

I added a `--rcfile` option and a default that will be passed to pylint,
as well as parsing and respecting the ignore setting while building the
file list.

I updated the contents of the readme to support the new features as well
as cleaning up some things like the broken badges.

Finally, I added a changelog.md for kicks and giggles.